Move curve structs to separate files
