+++ ValveResourceFormat/Resource/ResourceTypes/Choreo/Curves/BezierData.cs
@@ -1,0 +1,10 @@
+namespace ValveResourceFormat.ResourceTypes.Choreo.Curves
+{
+    public struct BezierData
+    {
+        public float InDegrees { get; set; }
+        public float InWeight { get; set; }
+        public float OutDegrees { get; set; }
+        public float OutWeight { get; set; }
+    }
+}
