+++ ValveResourceFormat/Resource/ResourceTypes/Choreo/Curves/CurveType.cs
@@ -1,0 +1,41 @@
+namespace ValveResourceFormat.ResourceTypes.Choreo.Curves
+{
+    public struct CurveType
+    {
+        private static readonly string[] Interpolators = [
+            "default",
+            "catmullrom_normalize_x",
+            "easein",
+            "easeout",
+            "easeinout",
+            "bspline",
+            "linear_interp",
+            "kochanek",
+            "kochanek_early",
+            "kochanek_late",
+            "simple_cubic",
+            "catmullrom",
+            "catmullrom_normalize",
+            "catmullrom_tangent",
+            "exponential_decay",
+            "hold",
+            "bezier",
+        ];
+        public byte InType { get; set; }
+        public byte OutType { get; set; }
+        public string InTypeName
+        {
+            get
+            {
+                return Interpolators[InType];
+            }
+        }
+        public string OutTypeName
+        {
+            get
+            {
+                return Interpolators[OutType];
+            }
+        }
+    }
+}
